Move fetcher package to pkg level
