+++ plugins/auth0/pkg/app/exec.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 
 	"github.com/alecthomas/kong"
-	"github.com/aserto-dev/ds-load/plugins/auth0/pkg/app/fetch"
+	"github.com/aserto-dev/ds-load/plugins/auth0/pkg/fetch"
 	"github.com/aserto-dev/ds-load/sdk/plugin"
 	"github.com/aserto-dev/ds-load/sdk/transform"
 )
