+++ plugins/azuread/pkg/fetch/fetch.go
